Move IAccountRepository from shared to domain
