+++ src/domain/repositories/account.repository.ts
@@ -1,0 +1,6 @@
+import { Account, AccountType } from '@/domain/entities';
+
+export interface IAccountRepository {
+  create(id: string, type: AccountType): Promise<Account>;
+  findById(id: string): Promise<Account | null>;
+}
