--- src/shared/types/account.ts
@@ -1,6 +1,0 @@
-import { Account, AccountType } from '@/domain/entities';
-
-export interface IAccountRepository {
-  create(id: string, type: AccountType): Promise<Account>;
-  findById(id: string): Promise<Account | null>;
-}
