+++ src/shared/types/index.ts
@@ -1,3 +1,2 @@
 export * from './response';
 export * from './database';
-export * from './account';
